% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.d.R
\name{mc.d}
\alias{mc.d}
\title{Monte Carlo Based Heterogeneity Test for Residual Variances in Random- or Mixed- Effects Model of Standardized Mean Differences (d)}
\usage{
mc.d(n1, n2, d, model = "random", mods = NULL, nrep = 10^4,
  p_cut = 0.05, mc.include = FALSE)
}
\arguments{
\item{n1}{a vector of sample sizes from group 1 in each of the included studies.}

\item{n2}{a vector of sample sizes from group 2 in each of the included studies.}

\item{d}{a vector of unbiased estimates of standardized mean differences.}

\item{model}{choice of random- or mixed- effects models. Can only be set to \code{"random"}, or \code{"mixed"}.}

\item{mods}{optional argument to include one or more moderators in the model. \code{mods} is NULL for random-effects model and a dataframe for mixed-effects model. A single moderator can be given as a vector of length \eqn{k} specifying the values of the moderator. Multiple moderators are specified by giving a matrix with \eqn{k} rows and as many columns as there are moderator variables. See \code{\link[metafor]{rma}} for more details.}

\item{nrep}{number of replications used in Monte Carlo Simulations. Default to 10^4.}

\item{p_cut}{cutoff for p-values, which is the alpha level. Default to 0.05.}

\item{mc.include}{If TRUE, Monte Carlo critical values will be included in the output.}
}
\description{
\code{mc.d} returns the Monte Carlo based tests of the residual heterogeneity in random- or mixed- effects model of standardized mean differences (d).
}
\details{
This function returns the test statistics as well as their p-value and significances using (1) Q-test, (2) Monte Carlo Based Heterogeneity Test with Maximum Likelihood (ML), and (3) Monte Carlo Based Heterogeneity Test with Restricted Maximum Likelihood (REML).

The results of significances are classified as "sig" or "n.s" based on the cutoff p-value (i.e., alpha level). "sig" means that the between-study heterogeneity is significantly different from zero wheras "n.s" means the between-study heterogeneity is not significantly different from zero. The default cutoff alpha level is 0.05
}
\examples{
# A meta-analysis of 18 studies in which the effect of open versus traditional education on students' self-concept was studied (Hedges et al., 1981).
data(selfconcept)
# n1 and n2 are lists of samples sizes in two groups
n1 <- selfconcept$n1
n2 <- selfconcept$n2
# g is a list of biased estimates of standardized mean differences in the meta-analytical study
g <- selfconcept$g
cm <- (1-3/(4*(n1+n2-2)-1)) #correct factor to compensate for small sample bias (Hedges, 1981)
d <- cm*g
\dontrun{
mc.run <- mc.d(n1, n2, d, model = 'random', p_cut = 0.05)
}

# A hypothetical meta-analysis of 15 studies with 3 moderators.
data(hypo_moder)
\dontrun{
mc.run2 <- mc.d(hypo_moder$n1, hypo_moder$n2, hypo_moder$d, model = 'mixed', mods = cbind(hypo_moder$cov.z1, hypo_moder$cov.z2, hypo_moder$cov.z1), p_cut = 0.05)
}
}
\references{
Hedges, L. V. (1981). Distribution theory for glass’s estimator of effect size and related estimators. Journal of Educational and Behavioral Statistics, 6(2), 107–128.

Hedges, L. V., Giaconia, R. M., & Gage, N. L. (1981). Meta-analysis of the effect of open and traditional instruction. Standford, CA: Stanford University, Program on Teaching Effectiveness.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1-48. URL: http://www.jstatsoft.org/v36/i03/
}
