% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.OR.R
\name{mc.OR}
\alias{mc.OR}
\title{Monte Carlo Based Heterogeneity Test for Residual Variances in Random- or Mixed- Effects Model of Log Odds Ratio (OR)}
\usage{
mc.OR(n, LOR, p_00, p_01, model = "random", mods = NULL, nrep = 10^4,
  p_cut = 0.05)
}
\arguments{
\item{n}{a vector of sample sizes in each of the included studies.}

\item{LOR}{a vector of log odds ratios in each of the included studies.}

\item{p_00}{a vector of probabilities of negatives on both Y1 and Y2.}

\item{p_01}{a vector of probabilities of negative on Y1 and positive on Y2.}

\item{model}{choice of random- or mixed- effects models. Can only be set to \code{"random"}, or \code{"mixed"}.}

\item{mods}{optional argument to include one or more moderators in the model. \code{mods} is NULL for random-effects model and a dataframe for mixed-effects model. A single moderator can be given as a vector of length \eqn{k} specifying the values of the moderator. Multiple moderators are specified by giving a matrix with \eqn{k} rows and as many columns as there are moderator variables. See \code{\link[metafor]{rma}} for details.}

\item{nrep}{number of replications used in Monte Carlo Simulations. Default to 10^4.}

\item{p_cut}{cutoff for p-values. Default to 0.05.}
}
\description{
\code{mc.d} returns the Monte Carlo based tests of the residual heterogeneity in random- or mixed- effects model of natural-logarithm-transformed observed odds ratio (OR).
}
\details{
This function returns the test statistics as well as their significances using (1) Q-test, (2) Monte Carlo Based Heterogeneity Test with Maximum Likelihood (ML), and (3) Monte Carlo Based Heterogeneity Test with Restricted Maximum Likelihood (REML).

The results of significances are classified as "sig" or "n.s" based on the cutoff p-value. "sig" means that the residual variance is significantly different from zero wheras "n.s" means the residual variance is not significantly different from zero.
}
\examples{
library(HSAUR2)
data(smoking)
# Total number of subjecs in each study
n <- smoking$tt + smoking$tc
# Y1: receive treatment; Y2: stop smoking
p_00 <- (smoking$tc - smoking$qc)/n  # not receive treatement yet not stop smoking
p_01 <- smoking$qc/n # not receive treatement but stop smoking
p_10 <- (smoking$tt - smoking$qt)/n # receive treatement but not stop smoking
p_11 <- smoking$qt/n # receive treatement and stop smoking
LOR <- log(p_11*p_00/p_01/p_10)
\dontrun{
mc.run <- mc.OR(n, LOR, p_00, p_01, model = 'random')
}
}
