% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.z.R
\name{mc.z}
\alias{mc.z}
\title{Monte Carlo Based Heterogeneity Test for Residual Variances in Random- or Mixed- Effects Model of Fisher's Transformed z Scores (z)}
\usage{
mc.z(n, z, model = "random", mods = NULL, nrep = 10^4,
  p_cut = 0.05)
}
\arguments{
\item{n}{a vector of sample sizes in each of the included studies.}

\item{z}{a vector of Fisher's transformed z scores (z).}

\item{model}{choice of random- or mixed- effects models. Can only be set to \code{"random"}, or \code{"mixed"}.}

\item{mods}{optional argument to include one or more moderators in the model. \code{mods} is NULL for random-effects model and a dataframe for mixed-effects model. A single moderator can be given as a vector of length \eqn{k} specifying the values of the moderator. Multiple moderators are specified by giving a matrix with \eqn{k} rows and as many columns as there are moderator variables. See \code{\link[metafor]{rma}} for details.}

\item{nrep}{number of replications used in Monte Carlo Simulations. Default to 10^4.}

\item{p_cut}{cutoff for p-values. Default to 0.05.}
}
\description{
\code{mc.d} returns the Monte Carlo based tests of the residual heterogeneity in random- or mixed- effects model of correlation coefficients transformed with Fisher's r-to-z transformation (z scores).
}
\details{
This function returns the test statistics as well as their significances using (1) Q-test, (2) Monte Carlo Based Heterogeneity Test with Maximum Likelihood (ML), and (3) Monte Carlo Based Heterogeneity Test with Restricted Maximum Likelihood (REML).

The results of significances are classified as "sig" or "n.s" based on the cutoff p-value. "sig" means that the residual variance is significantly different from zero wheras "n.s" means the residual variance is not significantly different from zero.
}
\examples{
# n is a list of samples sizes in the meta-analytical study.
n <- c(65, 30, 93, 36, 57, 30, 40, 13, 44, 58, 125, 10, 13)
# Pearson's Correction
r <- c(0.17, -0.45, -0.47, -0.13, -0.24, -0.15, -0.25, -0.66, -0.25, -0.23, -0.18, 0.18, -0.74)
# Fisher's Transformation
z <- 1/2*log((1+r)/(1-r))
\dontrun{
mc.run <- mc.z(n, z, model = 'random')
}
}
