% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.d.R
\name{mc.d}
\alias{mc.d}
\title{Monte Carlo Based Heterogeneity Test for Residual Variances in Random- or Mixed- Effects Model of Standardized Mean Differences (d)}
\usage{
mc.d(n1, n2, d, model = "random", mods = NULL, nrep = 10^4,
  p_cut = 0.05, mc.include = FALSE)
}
\arguments{
\item{n1}{a vector of sample sizes from group 1 in each of the included studies.}

\item{n2}{a vector of sample sizes from group 2 in each of the included studies.}

\item{d}{a vector of bias-corrected estimate of standardized mean differences (often reported as hedge's g)}

\item{model}{choice of random- or mixed- effects models. Can only be set to \code{"random"}, or \code{"mixed"}.}

\item{mods}{optional argument to include one or more moderators in the model. \code{mods} is NULL for random-effects model and a dataframe for mixed-effects model. A single moderator can be given as a vector of length \eqn{k} specifying the values of the moderator. Multiple moderators are specified by giving a matrix with \eqn{k} rows and as many columns as there are moderator variables. See \code{\link[metafor]{rma}} for details.}

\item{nrep}{number of replications used in Monte Carlo Simulations. Default to 10^4.}

\item{p_cut}{cutoff for p-values. Default to 0.05.}

\item{mc.include}{if Monte Carlo simulation results are included in the output.}
}
\description{
\code{mc.d} returns the Monte Carlo based tests of the residual heterogeneity in random- or mixed- effects model of standardized mean differences (d).
}
\details{
This function returns the test statistics as well as their p-value and significances using (1) Q-test, (2) Monte Carlo Based Heterogeneity Test with Maximum Likelihood (ML), and (3) Monte Carlo Based Heterogeneity Test with Restricted Maximum Likelihood (REML).

The results of significances are classified as "sig" or "n.s" based on the cutoff p-value. "sig" means that the residual variance is significantly different from zero wheras "n.s" means the residual variance is not significantly different from zero. The default cutoff p-value is 0.05
}
\examples{
# internal dataset for standardized mean differences (d)
data(exdat_d1)
# n1 and n2 are lists of samples sizes in two groups
n1 <- exdat_d1$n1
n2 <- exdat_d1$n2
# g is a list of standardized mean differences in the meta-analytical study
g <- exdat_d1$g
cm <- (1-3/(4*(n1+n2-2)-1)) #correct factor to compensate for small sample bias (Hedges & Olkin, 1985)
d <- cm*g
\dontrun{
mc.run <- mc.d(n1, n2, d, model = 'random', p_cut = 0.05)
}

# In the presense of moderators (cov.z1, cov.z2, cov.z3)
data(exdat_d2)
\dontrun{
mc.run2 <- mc.d(exdat_d2$n1, exdat_d2$n2, exdat_d2$d, model = 'mixed', mods = cbind(exdat_d2$cov.z1, exdat_d$cov.z2, exdat_d$cov.z1), p_cut = 0.05)
}
}
